+++ client/src/components/UI/main.ts
@@ -1,8 +1,10 @@
 import Input from "./Input/Input";
 import Button from "./button/Button";
 import Modal from "./modal/Modal";
+import Alert from "./alert/Alert";
 
 export const UI = {
+  Alert,
   Modal,
   Input,
   Button,