+++ client/src/pages/Home.tsx
@@ -1,15 +1,13 @@
 import React from "react";
-import { BrowserRouter } from "react-router-dom";
-// import { getUsers } from "./API/main";
 import Header from "../components/Header";
 import { UI } from "../components/UI/main";
 
-function HomePage() {
+function Home() {
   return (
     <div className="Home-page">
       <Header
         menuList={[
-          { url: "/i", label: "sigh in" },
+          { url: "/login", label: "sigh in" },
           { url: "/i", label: "transfer" },
         ]}
       />
@@ -17,4 +15,4 @@ function HomePage() {
   );
 }
 
-export default HomePage;
+export default Home;